Add hardware/google/graphics/zuma/libhwc2.1 to PRODUCT_SOONG_NAMESPACES
For zumapro it shares the same libacryl_plugins in zuma but for
libexynosdisplay it still has different version between zuma and
zumapro. In ag/30508311 it already split
hardware/google/graphics/zuma/libhwc2.1 to it's own soong_namespace
which need to add it to product config for zuma.

Bug: 357500180
Test: lunch aosp_husky-trunk_staging-userdebug; \
      m libexynosdisplay
Flag: EXEMPT refactor
Change-Id: I8adc95085067456aa2a29822e89f9777ece241f3