[AAPM] Enable MTE support for DevicePolicyManager for zuma
Set the system property that lets the DevicePolicyManager know it can
control Memory Tagging Extension on the device.

With this property set, when the user turns on AAPM, it will turn on
MTE.

This has been enabled for zumapro, now enabling for zuma devices.

Bug: 352420507
Test: Manual
Flag: android.security.aapm_feature_memory_tagging_extension
Change-Id: I14bfdd461076a499457c5a81ed151bfaebbfb2bb